% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_doi.R
\name{post_process_doi_citations}
\alias{post_process_doi_citations}
\alias{replace_resolved_doi_citations}
\title{Replace DOI citations in R Markdown document}
\usage{
post_process_doi_citations(input_file, bib)

replace_resolved_doi_citations()
}
\arguments{
\item{input_file}{Character. Path to the input file provided to the post-processor.}

\item{bib}{Character. A (vector of) path(s) to the BibTeX file(s).}
}
\value{
Returns `TRUE` invisibly.
}
\description{
This function reads an R Markdown document and replaces all DOI citations
with the corresponding entries from a BibTeX file. Requires the package
`bibtex` to be installed.
}
