% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verify_pandoc_version}
\alias{verify_pandoc_version}
\title{Verify minimum pandoc version}
\usage{
verify_pandoc_version(report = "error")
}
\arguments{
\item{report}{Character. In case pandoc version is < 2.0 errors if
\code{error}, warns if \code{warn}, or remains silent otherwise.}
}
\value{
Logical. \code{TRUE} if pandoc version is 2.0 or later, \code{FALSE}
  otherwise.
}
\description{
Verifies that pandoc version 2.0 or later is available.
}
\examples{

verify_pandoc_version(report = "silent")
}
