% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lua_filter.R
\name{add_lua_filter}
\alias{add_lua_filter}
\alias{add_wordcount_filter}
\title{Add Lua-filter to pandoc arguments}
\usage{
add_lua_filter(args, filter_name)

add_wordcount_filter(args)
}
\arguments{
\item{args}{Character. (Vector of) pandoc command line arguments.}

\item{filter_name}{Character. Name of the Lua-filter to add. See details.}
}
\value{
Character vector of pandoc command line arguments.

Character vector of pandoc command line arguments.
}
\description{
Adds a Lua-filter supplied by \pkg{rmdfiltr} to the vector of pandoc command
line arguments.

Adds a Lua-filter to count words to the vector of pandoc command line
arguments.
}
\details{
The following Lua-filters are available. Convenience functions named
  after the filter are available (e.g. \code{add_*_filter()}).

\describe{
  \item{\code{}}{}
}
}
\examples{

add_lua_filter(NULL, "wordcount")
add_wordcount_filter(NULL)
}
\concept{add_lua_filter}
